Update README.md
Changes have been made to how spreadsheets run the script. Might be best to include an image of the new menu to clarify.

Also in testing this I would say it's also good to provide a full 6 param example of the query and make sure to mention how it's the order of the params that indicate what that value does and as such you need to fill in the default or other options for the next one to work. I could be wrong though.